+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_email_validation_mode.php
@@ -3,5 +3,6 @@
 $container->loadFromExtension('framework', [
     'validation' => [
         'email_validation_mode' => 'html5-allow-no-tld',
+        'phone_number_validation_mode' => 'strict',
     ],
 ]);