+++ src/Symfony/Component/Validator/Constraints/PhoneNumberValidator.php
@@ -15,6 +15,8 @@ use libphonenumber\NumberParseException;
 use libphonenumber\PhoneNumberUtil;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
+use Symfony\Component\Validator\Exception\InvalidArgumentException;
+use Symfony\Component\Validator\Exception\LogicException;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
 use Symfony\Component\Validator\Exception\UnexpectedValueException;
 
@@ -23,6 +25,17 @@ use Symfony\Component\Validator\Exception\UnexpectedValueException;
  */
 final class PhoneNumberValidator extends ConstraintValidator
 {
+    private string $defaultMode;
+
+    public function __construct(string $defaultMode = PhoneNumber::MODE_E164)
+    {
+        if (!\in_array($defaultMode, PhoneNumber::VALIDATION_MODES, true)) {
+            throw new InvalidArgumentException('The "defaultMode" parameter value is not valid.');
+        }
+
+        $this->defaultMode = $defaultMode;
+    }
+
     public function validate(mixed $value, Constraint $constraint): void
     {
         if (!$constraint instanceof PhoneNumber) {
@@ -41,6 +54,18 @@ final class PhoneNumberValidator extends ConstraintValidator
 
         if (null !== $constraint->normalizer) {
             $value = ($constraint->normalizer)($value);
+        }
+
+        if (null === $constraint->mode) {
+            if (PhoneNumber::MODE_STRICT === $this->defaultMode && !class_exists(PhoneNumberUtil::class)) {
+                throw new LogicException(\sprintf('The "giggsey/libphonenumber-for-php-lite" library is required to make the "%s" constraint default to strict mode. Try running "composer require giggsey/libphonenumber-for-php-lite".', PhoneNumber::class));
+            }
+
+            $constraint->mode = $this->defaultMode;
+        }
+
+        if (!\in_array($constraint->mode, PhoneNumber::VALIDATION_MODES, true)) {
+            throw new InvalidArgumentException(\sprintf('The "%s::$mode" parameter value is not valid.', get_debug_type($constraint)));
         }
 
         if (!preg_match('/^\+[1-9]\d{1,14}$/D', $value)) {